requirements.txt
absl-py>=1.1.0
astunparse>=1.6.3
Bootstrap-Flask>=2.0.2
cachetools>=5.2.0
certifi>=2022.6.15
charset-normalizer>=2.0.12
click>=8.1.3
colorama>=0.4.5
dominate>=2.6.0
Flask>=2.1.2
flatbuffers>=1.12
gast>=0.4.0
google-api-core>=2.8.2
google-auth>=2.8.0
google-auth-oauthlib>=0.4.6
google-cloud-aiplatform>=1.14.0
google-cloud-bigquery>=2.34.4
google-cloud-core>=2.3.1
google-cloud-resource-manager>=1.5.1
google-cloud-storage>=2.4.0
google-crc32c>=1.3.0
google-pasta>=0.2.0
google-resumable-media>=2.3.3
googleapis-common-protos>=1.56.2
grpc-google-iam-v1>=0.12.4
grpcio>=1.46.3
grpcio-status>=1.46.3
gunicorn>=20.1.0
h5py>=3.7.0
idna>=3.3
itsdangerous>=2.1.2
Jinja2>=3.1.2
keras>=2.9.0
Keras-Preprocessing>=1.1.2
libclang>=14.0.1
Markdown>=3.3.7
MarkupSafe>=2.1.1
numpy>=1.22.4
oauthlib>=3.2.0
opencv-python>=4.6.0.66
opt-einsum>=3.3.0
packaging>=21.3
proto-plus>=1.20.6
protobuf>=3.19.4
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pyparsing>=3.0.9
python-dateutil>=2.8.2
python-dotenv>=0.20.0
requests>=2.28.0
requests-oauthlib>=1.3.1
rsa>=4.8
six>=1.16.0
tensorboard>=2.9.1
tensorboard-data-server>=0.6.1
tensorboard-plugin-wit>=1.8.1
tensorflow-estimator>=2.9.0
tensorflow-io-gcs-filesystem>=0.26.0
termcolor>=1.1.0
typing_extensions>=4.2.0
urllib3>=1.26.9
visitor>=0.1.3
Werkzeug>=2.1.2
wrapt>=1.14.1
WTForms>=3.0.1
